requirements.txt
mysqlclient
djangorestframework
django-filter
djangorestframework-jwt
djangorestframework-recursive